+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 0.1.0 - 2026-01-30
 
 - First usable version
+
+0.2.0
+
+- tzu uses command -r or --run-strat to run strategies, enabling the
+  addition of other commands in the future.